Extract drill cache to its own module
LocalStorage.js had grown a cluster of drill-specific state (storage
key, max-size constant, four CRUD-ish helpers, and a 14-line
comment) that doesn't belong in the user-session/preferences wrapper.

Moved to src/assorted/drillCache.js as named-function exports
(getDrillVocab, isDrillVocabEmpty, pushDrillVocab, clearDrillVocab).
LocalStorage drops the DrillVocab key, DRILL_MAX_PER_LANG constant,
and the four `LocalStorage.*` methods. Net: LocalStorage.js -65
lines, drillCache.js +61 lines, one self-contained concern in one
file.

Callers updated:
- App.js (startup seed)
- components/LoadingAnimation.js (mount gate)
- components/WaitDrill.js (queue build)
- exercises/ExerciseSession.js (exercise tee)
- reader/InteractiveText.js (translation-tap tee)
