+++ src/exercises/ExerciseSession.js
@@ -9,6 +9,7 @@ import SessionStorage from "../assorted/SessionStorage";
 import { CORRECT, SOLUTION, WRONG } from "./ExerciseConstants";
 import { EXERCISE_TYPES } from "./ExerciseTypeConstants";
 import LocalStorage from "../assorted/LocalStorage";
+import { pushDrillVocab } from "../assorted/drillCache";
 import { assignBookmarksToExercises } from "./assignBookmarksToExercises";
 import NextNavigation from "./exerciseTypes/NextNavigation";
 import { SpeechContext } from "../contexts/SpeechContext";
@@ -337,7 +338,7 @@ export default function ExerciseSession({ articleID, backButtonAction, toSchedul
 
   function teeBookmarkToDrillCache(b) {
     if (!b || !b.from || !b.to) return;
-    LocalStorage.pushDrillVocab(
+    pushDrillVocab(
       userDetails?.learned_language,
       [{ o: b.from, t: b.to }],
       "exercise",
